refactor types to data-structures
