+++ src/fed_rag/fl_tasks/pytorch.py
@@ -17,6 +17,7 @@ from torch.utils.data import DataLoader
 from typing_extensions import Self
 
 from fed_rag.base.fl_task import BaseFLTask, BaseFLTaskConfig
+from fed_rag.data_structures import TestResult, TrainResult
 from fed_rag.exceptions import (
     MissingRequiredNetParam,
     MissingTesterSpec,
@@ -27,7 +28,6 @@ from fed_rag.inspectors.pytorch import (
     TesterSignatureSpec,
     TrainerSignatureSpec,
 )
-from fed_rag.types import TestResult, TrainResult
 
 
 class PyTorchFLTaskConfig(BaseFLTaskConfig):
