+++ src/fed_rag/inspectors/pytorch/tester.py
@@ -3,13 +3,13 @@
 import inspect
 from typing import Any, Callable
 
+from fed_rag.data_structures import TestResult
 from fed_rag.exceptions import (
     InvalidReturnType,
     MissingDataParam,
     MissingNetParam,
 )
 from fed_rag.inspectors.common import TesterSignatureSpec
-from fed_rag.types import TestResult
 
 
 def inspect_tester_signature(fn: Callable) -> TesterSignatureSpec:
@@ -18,7 +18,7 @@ def inspect_tester_signature(fn: Callable) -> TesterSignatureSpec:
     # validate return type
     return_type = sig.return_annotation
     if (return_type is Any) or not issubclass(return_type, TestResult):
-        msg = "Tester should return a fed_rag.types.TestResult or a subclsas of it."
+        msg = "Tester should return a fed_rag.data_structures.TestResult or a subclsas of it."
         raise InvalidReturnType(msg)
 
     # inspect fn params
