+++ src/fed_rag/inspectors/pytorch/trainer.py
@@ -3,6 +3,7 @@
 import inspect
 from typing import Any, Callable
 
+from fed_rag.data_structures import TrainResult
 from fed_rag.exceptions import (
     InvalidReturnType,
     MissingDataParam,
@@ -10,7 +11,6 @@ from fed_rag.exceptions import (
     MissingNetParam,
 )
 from fed_rag.inspectors.common import TrainerSignatureSpec
-from fed_rag.types import TrainResult
 
 
 def inspect_trainer_signature(fn: Callable) -> TrainerSignatureSpec:
@@ -19,7 +19,7 @@ def inspect_trainer_signature(fn: Callable) -> TrainerSignatureSpec:
     # validate return type
     return_type = sig.return_annotation
     if (return_type is Any) or not issubclass(return_type, TrainResult):
-        msg = "Trainer should return a fed_rag.types.TrainResult or a subclsas of it."
+        msg = "Trainer should return a fed_rag.data_structures.TrainResult or a subclsas of it."
         raise InvalidReturnType(msg)
 
     # inspect fn params
