+++ src/fed_rag/trainer_managers/pytorch.py
@@ -8,13 +8,13 @@ from torch.utils.data import DataLoader
 from typing_extensions import assert_never
 
 from fed_rag.base.trainer_manager import BaseRAGTrainerManager
+from fed_rag.data_structures.results import TestResult, TrainResult
 from fed_rag.decorators import federate
 from fed_rag.exceptions.trainer_manager import (
     UnspecifiedGeneratorTrainer,
     UnspecifiedRetrieverTrainer,
 )
 from fed_rag.fl_tasks.pytorch import PyTorchFLTask
-from fed_rag.types.results import TestResult, TrainResult
 
 
 class TrainingArgs(BaseModel):
