+++ src/fed_rag/trainers/huggingface/lsr.py
@@ -8,6 +8,7 @@ from pydantic import PrivateAttr, model_validator
 from fed_rag import RAGSystem
 from fed_rag.base.trainer import BaseRetrieverTrainer
 from fed_rag.data_collators.huggingface import DataCollatorForLSR
+from fed_rag.data_structures.results import TestResult, TrainResult
 from fed_rag.exceptions import (
     InvalidDataCollatorError,
     InvalidLossError,
@@ -17,7 +18,6 @@ from fed_rag.exceptions import (
 )
 from fed_rag.loss.pytorch.lsr import LSRLoss
 from fed_rag.trainers.huggingface.mixin import HuggingFaceTrainerMixin
-from fed_rag.types.results import TestResult, TrainResult
 from fed_rag.utils.huggingface import _validate_rag_system
 
 try:
