+++ src/fed_rag/types/bridge.py
@@ -1,12 +1,19 @@
-"""Bridge type definitions for fed-rag."""
+"""Bridge type definitions for fed-rag.
 
-from typing import TypedDict
+Note: The BridgeMetadata implementation has moved to fed_rag.data_structures.bridge.
+This module is maintained for backward compatibility.
+"""
 
+import warnings
 
-class BridgeMetadata(TypedDict):
-    """Type definition for bridge metadata."""
+from ..data_structures.bridge import BridgeMetadata
 
-    bridge_version: str
-    framework: str
-    compatible_versions: list[str]
-    method_name: str
+warnings.warn(
+    "Importing BridgeMetadata from fed_rag.types.bridge is deprecated and will be "
+    "removed in a future release. Use fed_rag.data_structures.bridge or "
+    "fed_rag.data_structures instead.",
+    DeprecationWarning,
+    stacklevel=2,  # point to users import statement
+)
+
+__all__ = ["BridgeMetadata"]
