+++ tests/bridges/conftest.py
@@ -10,8 +10,8 @@ from fed_rag.base.generator import BaseGenerator
 from fed_rag.base.retriever import BaseRetriever
 from fed_rag.base.tokenizer import BaseTokenizer
 from fed_rag.core.rag_system import RAGSystem
+from fed_rag.data_structures import RAGConfig
 from fed_rag.knowledge_stores.in_memory import InMemoryKnowledgeStore
-from fed_rag.types import RAGConfig
 
 
 class _TestDataset(Dataset):
