+++ tests/data_collators/huggingface/test_data_collator_for_ralt.py
@@ -14,6 +14,7 @@ from fed_rag.data_collators.huggingface.ralt import (
     DEFAULT_EXAMPLE_TEMPLATE,
     DataCollatorForRALT,
 )
+from fed_rag.data_structures import KnowledgeNode, SourceNode
 from fed_rag.exceptions import (
     DataCollatorError,
     FedRAGError,
@@ -23,7 +24,6 @@ from fed_rag.generators.huggingface import HFPeftModelGenerator
 from fed_rag.retrievers.huggingface.hf_sentence_transformer import (
     HFSentenceTransformerRetriever,
 )
-from fed_rag.types import KnowledgeNode, SourceNode
 
 
 def test_huggingface_extra_missing(mock_rag_system: RAGSystem) -> None:
