+++ tests/knowledge_stores/test_in_memory_managed.py
@@ -5,9 +5,9 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from fed_rag.base.knowledge_store import BaseKnowledgeStore
+from fed_rag.data_structures.knowledge_node import KnowledgeNode
 from fed_rag.exceptions import KnowledgeStoreNotFoundError
 from fed_rag.knowledge_stores.in_memory import ManagedInMemoryKnowledgeStore
-from fed_rag.types.knowledge_node import KnowledgeNode
 
 
 @pytest.fixture
