+++ tests/knowledge_stores/test_knowledge_node.py
@@ -2,10 +2,10 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from fed_rag.types.knowledge_node import KnowledgeNode, NodeContent
+from fed_rag.data_structures.knowledge_node import KnowledgeNode, NodeContent
 
 
-@patch("fed_rag.types.knowledge_node.uuid")
+@patch("fed_rag.data_structures.knowledge_node.uuid")
 def test_text_knowledge_node_init(mock_uuid: MagicMock) -> None:
     mock_uuid.uuid4.return_value = "mock_id"
     node = KnowledgeNode(
@@ -27,7 +27,7 @@ def test_text_knowledge_node_init_raises_validation_error() -> None:
         )
 
 
-@patch("fed_rag.types.knowledge_node.uuid")
+@patch("fed_rag.data_structures.knowledge_node.uuid")
 def test_image_knowledge_node_init(mock_uuid: MagicMock) -> None:
     mock_uuid.uuid4.return_value = "mock_id"
     node = KnowledgeNode(
@@ -52,7 +52,7 @@ def test_image_knowledge_node_init_raises_validation_error() -> None:
         )
 
 
-@patch("fed_rag.types.knowledge_node.uuid")
+@patch("fed_rag.data_structures.knowledge_node.uuid")
 def test_multimodal_knowledge_node_init(mock_uuid: MagicMock) -> None:
     mock_uuid.uuid4.return_value = "mock_id"
     node = KnowledgeNode(
