+++ tests/trainers/conftest.py
@@ -9,8 +9,8 @@ from fed_rag.base.generator import BaseGenerator
 from fed_rag.base.retriever import BaseRetriever
 from fed_rag.base.tokenizer import BaseTokenizer
 from fed_rag.base.trainer import BaseRetrieverTrainer, BaseTrainer
+from fed_rag.data_structures.results import TestResult, TrainResult
 from fed_rag.knowledge_stores.in_memory import InMemoryKnowledgeStore
-from fed_rag.types.results import TestResult, TrainResult
 
 
 class MockRetriever(BaseRetriever):
